Bug:Fix tournament bugs
Fix Error Scroll-to-Top Issue
Ensure that when an error occurs, the page automatically scrolls to the top so the user notices it.

Change URL for Searching Venues (TO)
Update the URL endpoint used to search for venues.

Change URL for Updating Match Score (TO)
Update the URL endpoint used to update the match score.

Fix Validation Issue in Booking Dates
Ensure that the end date is not earlier than the start date in the booking form.